Widen the EGC embedding accumulator to int64
synthesize summed one EGC's constituent embedding rows into an int32.
The design assumed a cluster is at most a few dozen codepoints, but
nothing enforces that and the cluster comes from caller input: a long
run of combining marks is a single EGC of unbounded length. A ~400KB
line of them overflowed the accumulator, which is UB (UBSan flags it;
release builds silently score the boundary wrong).

Values that fit in int32 accumulate identically, so the bit-exactness
contract with the trainer's reference forward is unchanged, and the
throughput is the same within noise (this is the table-build and
multi-codepoint fallback path, not the inference hot loop).

+++ src/mlp/precompute.cpp
@@ -18,11 +18,17 @@ namespace segmentlib::mlp {
 template <class Emit>
 void PrecomputeTable::synthesize(std::span<const std::uint32_t> rows,
                                  std::size_t j, Emit emit) const {
-    std::int32_t sum_c[256];  // the loader rejects embed_dim > 256
+    // int64, not int32: `rows` is one EGC's constituent codepoints, and an EGC
+    // is caller-controlled — a long run of combining marks is a single cluster
+    // of unbounded length. At int32 a ~400 KB line of them overflowed the
+    // accumulator (UB), so the width is a correctness requirement, not
+    // headroom. Values that fit in int32 accumulate identically here, so the
+    // bit-exactness contract with the trainer's reference forward is unchanged.
+    std::int64_t sum_c[256];  // the loader rejects embed_dim > 256
     const std::size_t d = d_;
     assert(d <= std::size(sum_c));
     for (std::size_t k = 0; k < d; ++k) {
-        std::int32_t s = 0;
+        std::int64_t s = 0;
         for (const std::uint32_t row : rows) {
             s += emb_q_[static_cast<std::size_t>(row) * d + k];
         }
